+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.kt
@@ -310,7 +310,9 @@ class SkyHanniMod {
         loadModule(VoltHighlighter())
         loadModule(RiftOdonata())
         loadModule(RiftAgaricusCap())
+        loadModule(KloonHacking())
         loadModule(EnigmaSoulWaypoints)
+        loadModule(DungeonLividFinder)
         loadModule(CruxTalismanDisplay())
 
         init()